pythonProject/requirements.txt
paho-mqtt~=1.6.1
python-telegram-bot~=13.11